refactor: 숫자 validator 리팩토링
RandomNumberGenerator에 종속적으로 있던 숫자 유효성 검사기능을 클래스로 분리하였다.
